Apply code suggestions from Sekhar: use SVG images instead of PNGs and organize auxiliary build components into a separate file
