+++ client/src/features/sessionsV2/SessionList/SessionItem.tsx
@@ -23,7 +23,7 @@ import { Project } from "../../projectsV2/api/projectV2.api";
 import { getShowSessionUrlByProject } from "../SessionsV2";
 import StartSessionButton from "../StartSessionButton";
 import type { SessionLauncher } from "../api/sessionLaunchersV2.api";
-import { ActiveSessionButton } from "../components/SessionButton/ActiveSessionButton";
+import ActiveSessionButton from "../components/SessionButton/ActiveSessionButton";
 import {
   SessionBadge,
   SessionStatusV2Description,
