+++ client/src/features/sessionsV2/SessionStyles.constants.ts
@@ -1,8 +1,26 @@
-import linePlaying from "../../styles/assets/linePlaying.png";
-import lineFailed from "../../styles/assets/lineFailed.png";
-import linePaused from "../../styles/assets/linePaused.png";
-import lineStopped from "../../styles/assets/lineStopped.png";
-import lineBlock from "../../styles/assets/lineBlock.png";
+/*!
+ * Copyright 2025 - Swiss Data Science Center (SDSC)
+ * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+ * Eidgenössische Technische Hochschule Zürich (ETHZ).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import linePlaying from "../../styles/assets/linePlaying.svg";
+import lineFailed from "../../styles/assets/lineFailed.svg";
+import linePaused from "../../styles/assets/linePaused.svg";
+import lineStopped from "../../styles/assets/lineStopped.svg";
+import lineBlock from "../../styles/assets/lineBlock.svg";
 
 export const SESSION_STATES = {
   RUNNING: "running",
